cleanup some APIs
