+++ crates/wasm/src/main.rs
@@ -1,13 +1,9 @@
-mod loader;
-mod lowering;
-
-use cairo_m_common::InputValue;
 use cairo_m_compiler_codegen::compile_module;
-use cairo_m_compiler_mir::{PassManager, PrettyPrint};
-use cairo_m_runner::run_cairo_program;
+use cairo_m_compiler_mir::PassManager;
+use cairo_m_wasm::loader::{BlocklessDagModule, WasmLoadError};
+use cairo_m_wasm::lowering::lower_program_to_mir;
 use clap::Parser;
-use loader::{BlocklessDagModule, WasmLoadError};
-use lowering::lower_program_to_mir;
+use std::process;
 use std::{fs, path::PathBuf};
 use tracing::Level;
 
@@ -30,10 +26,6 @@ struct Args {
     /// Function name to run after compilation (entrypoint)
     #[arg(short = 'f', long)]
     function: Option<String>,
-
-    /// Arguments to pass to the entrypoint (repeat -a for multiple args)
-    #[arg(short = 'a', long = "arg")]
-    args: Vec<i64>,
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -47,38 +39,29 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mir_module = lower_program_to_mir(&module, PassManager::standard_pipeline())?;
     let program = compile_module(&mir_module)?;
 
-    // If a function is provided, execute it with args and print return values
-    if let Some(func) = args.function.as_deref() {
-        if args.verbose {
+    let json = sonic_rs::to_string_pretty(&program).unwrap_or_else(|e| {
+        eprintln!("Failed to serialize program: {}", e);
+        process::exit(1);
+    });
+
+    // Write output or print to stdout
+    match args.output {
+        Some(output_path) => {
+            fs::write(&output_path, &json).unwrap_or_else(|e| {
+                eprintln!(
+                    "Failed to write output file '{}': {}",
+                    output_path.display(),
+                    e
+                );
+                process::exit(1);
+            });
             println!(
-                "Running entrypoint '{}' with {} args",
-                func,
-                args.args.len()
+                "Compilation successful. Output written to '{}'",
+                output_path.display()
             );
         }
-        let input_values = args
-            .args
-            .iter()
-            .map(|&v| InputValue::Number(v))
-            .collect::<Vec<_>>();
-        let output = run_cairo_program(&program, func, &input_values, Default::default())?;
-        println!("{:?}", output.return_values);
-    } else {
-        // Otherwise, serialize the program to JSON
-        let json = sonic_rs::to_string_pretty(&program)?;
-
-        // Write output or print to stdout
-        match args.output {
-            Some(output_path) => {
-                fs::write(&output_path, &json)?;
-                println!(
-                    "Compilation successful. Output written to '{}'",
-                    output_path.display()
-                );
-            }
-            None => {
-                println!("{}", json);
-            }
+        None => {
+            println!("{}", json);
         }
     }
